Updated red-black tree for pointer size

+++ rbtree.h
@@ -38,11 +38,11 @@
 
 #define RB_EMPTY_ROOT(root)  ((root)->rb_node == 0)
 
-/* 'empty' nodes are nodes that are known not to be inserted in an rbree */
+/* 'empty' nodes are nodes that are known not to be inserted in an rbtree */
 #define RB_EMPTY_NODE(node)  \
-	((node)->__rb_parent_color == (unsigned long)(node))
+	((node)->__rb_parent_color == (uintptr_t)(node))
 #define RB_CLEAR_NODE(node)  \
-	((node)->__rb_parent_color = (unsigned long)(node))
+	((node)->__rb_parent_color = (uintptr_t)(node))
 
 
 extern void rb_insert_color(struct rb_node *, struct rb_root *);
@@ -62,7 +62,7 @@ extern void rb_replace_node(struct rb_node *victim, struct rb_node *new,
 static inline void rb_link_node(struct rb_node * node, struct rb_node * parent,
 				struct rb_node ** rb_link)
 {
-	node->__rb_parent_color = (unsigned long)parent;
+	node->__rb_parent_color = (uintptr_t)parent;
 	node->rb_left = node->rb_right = 0;
 
 	*rb_link = node;
